@@ -26,15 +26,17 @@
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "~7.0.0",
-        "@angular/common": "~7.0.0",
-        "@angular/compiler": "~7.0.0",
-        "@angular/core": "~7.0.0",
-        "@angular/forms": "~7.0.0",
-        "@angular/http": "~7.0.0",
-        "@angular/platform-browser": "~7.0.0",
-        "@angular/platform-browser-dynamic": "~7.0.0",
-        "@angular/router": "~7.0.0",
+        "@angular/animations": "^7.1.1",
+        "@angular/cdk": "^7.1.0",
+        "@angular/common": "~7.1.1",
+        "@angular/compiler": "~7.1.1",
+        "@angular/core": "~7.1.1",
+        "@angular/forms": "~7.1.1",
+        "@angular/http": "~7.1.1",
+        "@angular/material": "^7.1.0",
+        "@angular/platform-browser": "~7.1.1",
+        "@angular/platform-browser-dynamic": "~7.1.1",
+        "@angular/router": "~7.1.1",
         "@ngrx/effects": "^6.1.2",
         "@ngrx/entity": "^6.1.2",
         "@ngrx/router-store": "^6.1.2",
@@ -42,8 +44,9 @@
         "@ngrx/store-devtools": "^6.1.2",
         "@swimlane/ngx-datatable": "^14.0.0",
         "core-js": "^2.5.4",
-        "material-icons": "^0.2.3",
-        "materialize-css": "^1.0.0",
+        "hamburgers": "^1.1.3",
+        "hammerjs": "^2.0.8",
+        "material-design-icons": "^3.0.1",
         "ng2-charts": "^1.6.0",
         "rxjs": "~6.3.3",
         "zone.js": "~0.8.26"
@@ -51,8 +54,8 @@
     "devDependencies": {
         "@angular-devkit/build-angular": "~0.10.0",
         "@angular/cli": "~7.0.6",
-        "@angular/compiler-cli": "~7.0.0",
-        "@angular/language-service": "~7.0.0",
+        "@angular/compiler-cli": "~7.1.1",
+        "@angular/language-service": "~7.1.1",
         "@compodoc/compodoc": "^1.1.5",
         "@ngrx/schematics": "^6.1.2",
         "@types/jasmine": "~2.8.8",
@@ -62,6 +65,7 @@
         "@types/node": "~8.9.4",
         "codelyzer": "~4.5.0",
         "cross-var": "^1.1.0",
+        "include-media": "^1.4.9",
         "jasmine-core": "~2.99.1",
         "jasmine-spec-reporter": "~4.2.1",
         "karma": "~3.0.0",
