@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
+import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
 import { HomeComponent } from './core/home/home.component';
 import { NotFoundComponent } from './core/not-found/not-found.component';
 import { AdminGuard } from './core/guard/admin.guard';
@@ -7,9 +7,8 @@ import { SignInComponent } from './auth/sign-in/sign-in.component';
 
 
 const routes:Routes = [
-    { path:'', pathMatch:'full', redirectTo:'home' },
+    { path:'', component:HomeComponent },
     { path:'admin', loadChildren:'./admin/admin.module#AdminModule', canLoad:[ AdminGuard ] },
-    { path:'home', component:HomeComponent },
     { path:'signin', component:SignInComponent },
     { path:'not-found', component:NotFoundComponent },
     { path:'**', component:NotFoundComponent }
@@ -18,8 +17,9 @@ const routes:Routes = [
 
 @NgModule({
     imports:[ RouterModule.forRoot(routes, {
+        preloadingStrategy:PreloadAllModules,
         scrollPositionRestoration:'enabled',
-        anchorScrolling:'enabled'
+        anchorScrolling:'enabled',
     }) ],
     exports:[ RouterModule ]
 })
